Fixed some 64 bit specific bugs
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/FreeMat4@3525 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ libs/libMath/libLAPACK_C/f2c.h
@@ -38,9 +38,9 @@ typedef short flag;
 typedef short ftnlen;
 typedef short ftnint;
 #else
-typedef long int flag;
-typedef long int ftnlen;
-typedef long int ftnint;
+typedef int flag;
+typedef int ftnlen;
+typedef int ftnint;
 #endif
 
 /*external read, write*/
